# docs/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "学科知识整合智能体",
    "description": "学科知识整合智能体 API：多格式解析、统一 JSON、证据链和任务状态接口。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/uploads/sessions": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Create Upload Session",
        "operationId": "create_upload_session_api_uploads_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadSessionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadSessionRecord"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/uploads/sessions/{session_id}": {
      "get": {
        "tags": [
          "uploads"
        ],
        "summary": "Get Upload Session",
        "operationId": "get_upload_session_api_uploads_sessions__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadSessionRecord"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/uploads/sessions/{session_id}/chunks/{chunk_index}": {
      "put": {
        "tags": [
          "uploads"
        ],
        "summary": "Upload Chunk",
        "operationId": "upload_chunk_api_uploads_sessions__session_id__chunks__chunk_index__put",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "chunk_index",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Chunk Index"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_chunk_api_uploads_sessions__session_id__chunks__chunk_index__put"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadChunkResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/uploads/sessions/{session_id}/complete": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Complete Upload Session",
        "operationId": "complete_upload_session_api_uploads_sessions__session_id__complete_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadSessionCompleteResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/uploads/sessions/{session_id}/complete-async": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Complete Upload Session Async",
        "operationId": "complete_upload_session_async_api_uploads_sessions__session_id__complete_async_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AsyncTextbookParseResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message",
          "code"
        ],
        "title": "ApiErrorResponse"
      },
      "AsyncTextbookParseResponse": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/JobRecord"
          },
          "accepted": {
            "type": "boolean",
            "title": "Accepted",
            "default": true
          },
          "upload_session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upload Session Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job"
        ],
        "title": "AsyncTextbookParseResponse"
      },
      "Body_upload_chunk_api_uploads_sessions__session_id__chunks__chunk_index__put": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_chunk_api_uploads_sessions__session_id__chunks__chunk_index__put"
      },
...
      "Chunk": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "raw_file_id": {
            "type": "string",
            "title": "Raw File Id"
          },
          "section_id": {
            "type": "string",
            "title": "Section Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "order_index": {
            "type": "integer",
            "title": "Order Index"
          },
          "char_start": {
            "type": "integer",
            "title": "Char Start"
          },
          "char_end": {
            "type": "integer",
            "title": "Char End"
          },
          "char_count": {
            "type": "integer",
            "title": "Char Count"
          },
          "source_locator": {
            "$ref": "#/components/schemas/SourceLocator"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "raw_file_id",
          "section_id",
          "text",
          "order_index",
          "char_start",
          "char_end",
          "char_count",
          "source_locator"
        ],
        "title": "Chunk"
      },
      "DocumentElement": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "raw_file_id": {
            "type": "string",
            "title": "Raw File Id"
          },
          "type": {
            "$ref": "#/components/schemas/DocumentElementType"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "order_index": {
            "type": "integer",
            "title": "Order Index"
          },
          "parent_section_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Section Id"
          },
          "source_locator": {
            "$ref": "#/components/schemas/SourceLocator"
          },
          "char_count": {
            "type": "integer",
            "title": "Char Count"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "raw_file_id",
          "type",
          "text",
          "order_index",
          "source_locator",
          "char_count"
        ],
        "title": "DocumentElement"
      },
...
      "JobRecord": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "job_type": {
            "$ref": "#/components/schemas/JobType"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "progress": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/PipelineStepRecord"
            },
            "type": "array",
            "title": "Steps"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable",
            "default": false
          },
          "context_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Path"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "job_type",
          "status"
        ],
        "title": "JobRecord"
      },
...
      "JobStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "completed",
          "failed"
        ],
        "title": "JobStatus"
      },
      "JobType": {
        "type": "string",
        "enum": [
          "textbook_upload",
          "textbook_batch_upload",
          "textbook_parse",
          "textbook_pipeline",
          "large_file_upload",
          "graph_build",
          "converted_textbook_import"
        ],
        "title": "JobType"
      },
...
      "ParsedTextbook": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "raw_file": {
            "$ref": "#/components/schemas/RawFile"
          },
          "elements": {
            "items": {
              "$ref": "#/components/schemas/DocumentElement"
            },
            "type": "array",
            "title": "Elements"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/Section"
            },
            "type": "array",
            "title": "Sections"
          },
          "chunks": {
            "items": {
              "$ref": "#/components/schemas/Chunk"
            },
            "type": "array",
            "title": "Chunks"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "raw_file",
          "elements",
          "sections",
          "chunks"
        ],
        "title": "ParsedTextbook"
      },
      "PipelineStepName": {
        "type": "string",
        "enum": [
          "assemble_file",
          "detect_format",
          "parse_elements",
          "build_sections",
          "chunk_sections",
          "persist_parsed"
        ],
        "title": "PipelineStepName"
      },
      "PipelineStepRecord": {
        "properties": {
          "name": {
            "$ref": "#/components/schemas/PipelineStepName"
          },
          "status": {
            "$ref": "#/components/schemas/PipelineStepStatus",
            "default": "queued"
          },
          "progress": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "PipelineStepRecord"
      },
      "PipelineStepStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "completed",
          "failed",
          "skipped"
        ],
        "title": "PipelineStepStatus"
      },
      "RawFile": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "format": {
            "type": "string",
            "title": "Format"
          },
          "source_type": {
            "type": "string",
            "enum": [
              "uploaded",
              "converted_textbook",
              "generated"
            ],
            "title": "Source Type",
            "default": "uploaded"
          },
          "storage_path": {
            "type": "string",
            "title": "Storage Path"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "text_char_count": {
            "type": "integer",
            "title": "Text Char Count",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "original_filename",
          "title",
          "format",
          "storage_path",
          "sha256",
          "size_bytes"
        ],
        "title": "RawFile"
      },
...
      "TextbookUploadResponse": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/JobRecord"
          },
          "raw_file_id": {
            "type": "string",
            "title": "Raw File Id"
          },
          "parsed_output_path": {
            "type": "string",
            "title": "Parsed Output Path"
          },
          "parsed_textbook": {
            "$ref": "#/components/schemas/ParsedTextbook"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job",
          "raw_file_id",
          "parsed_output_path",
          "parsed_textbook"
        ],
        "title": "TextbookUploadResponse"
      },
      "UploadChunkResponse": {
        "properties": {
          "session": {
            "$ref": "#/components/schemas/UploadSessionRecord"
          },
          "chunk_index": {
            "type": "integer",
            "title": "Chunk Index"
          },
          "received_bytes": {
            "type": "integer",
            "title": "Received Bytes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session",
          "chunk_index",
          "received_bytes"
        ],
        "title": "UploadChunkResponse"
      },
      "UploadSessionCompleteResponse": {
        "properties": {
          "session": {
            "$ref": "#/components/schemas/UploadSessionRecord"
          },
          "job": {
            "$ref": "#/components/schemas/JobRecord"
          },
          "parsed_upload": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TextbookUploadResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session",
          "job"
        ],
        "title": "UploadSessionCompleteResponse"
      },
      "UploadSessionCreateRequest": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "total_size_bytes": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Total Size Bytes"
          },
          "total_chunks": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Total Chunks"
          },
          "chunk_size_bytes": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Chunk Size Bytes"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "parse_on_complete": {
            "type": "boolean",
            "title": "Parse On Complete",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "filename",
          "total_size_bytes",
          "total_chunks",
          "chunk_size_bytes"
        ],
        "title": "UploadSessionCreateRequest"
      },
      "UploadSessionRecord": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "total_size_bytes": {
            "type": "integer",
            "title": "Total Size Bytes"
          },
          "total_chunks": {
            "type": "integer",
            "title": "Total Chunks"
          },
          "chunk_size_bytes": {
            "type": "integer",
            "title": "Chunk Size Bytes"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "parse_on_complete": {
            "type": "boolean",
            "title": "Parse On Complete",
            "default": true
          },
          "status": {
            "$ref": "#/components/schemas/UploadSessionStatus",
            "default": "created"
          },
          "uploaded_chunks": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Uploaded Chunks"
          },
          "missing_chunks": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Missing Chunks"
          },
          "received_bytes": {
            "type": "integer",
            "title": "Received Bytes",
            "default": 0
          },
          "upload_progress": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Upload Progress",
            "default": 0
          },
          "parse_progress": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Parse Progress",
            "default": 0
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "assembled_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assembled Path"
          },
          "raw_file_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw File Id"
          },
          "parsed_output_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parsed Output Path"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "filename",
          "total_size_bytes",
          "total_chunks",
          "chunk_size_bytes"
        ],
        "title": "UploadSessionRecord"
      },
      "UploadSessionStatus": {
        "type": "string",
        "enum": [
          "created",
          "uploading",
          "assembling",
          "parsing",
          "completed",
          "failed"
        ],
        "title": "UploadSessionStatus"
      }
    }
  }
}